requirements.txt
blinker==1.4
cachetools==2.1.0
certifi==2018.11.29
Click==7.0
decorator==4.3.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Moment==0.6.0
Flask-WTF==0.14.2
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
httplib2==0.12.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mkl-fft==1.0.10
mkl-random==1.0.2
networkx==2.2
numpy==1.15.4
pandas==0.23.4
pyasn1==0.4.4
pyasn1-modules==0.2.3
PyScaffold==3.1
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.9
rsa==3.4.2
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
